refactor: Split Component into another file
